# Conflicts:
#	dev/src/main/java/DataAccessLayer/SuperLi.db

+++ dev/src/main/java/PresentationLayer/Transport/IOTransport.java
@@ -198,11 +198,14 @@ public class IOTransport {
             System.out.println("Please choose the transport ID you wish to remove:");
             int transportToDelete = integerParse(scanner.nextLine());
             //try to delete the given transport, if returns false the id is not in the system
-            boolean deleted = facadeController.DeleteTransportFronDB(transportToDelete);
-            if (deleted) {
+            if (facadeController.checkIfTransportExist(transportToDelete)) {
                 facadeController.removeDatesFromDriverAndTruck(transportToDelete);
-                System.out.println("The transport deleted successfully.\n");
-            } else {
+                boolean deleted = facadeController.DeleteTransportFronDB(transportToDelete);
+                if (deleted) {
+                    System.out.println("The transport deleted successfully.\n");
+                }
+            }
+            else {
                 System.out.println("The transport ID:" + transportToDelete + " is not in the system, operation canceled.\n");
             }
         }
